+++ ReleaseNotes.md
@@ -11,6 +11,7 @@ Here's a list of the latest changes:
 - [Change - Do not display 404 when session expires. Instead we display a more relevant page.](https://trello.com/c/0eopf73d/553-404-is-not-a-nice-page-to-display-when-session-is-no-longer-available)
 - [Change - Making the add-ticket button really big on the booking page](https://trello.com/c/MyVxAkUq/529-add-a-ticket-very-easy-to-miss-bem)
 - [Change - Removing the ticketing close date - Bem's suggestion](https://trello.com/c/XDbm6zgp/531-remove-ticketing-closing-date-from-booking-move-to-ticket-editing-as-part-of-next-release)
+- [Change - Replacing the word ad with Event for the new booking notification](https://trello.com/c/NjIbtkk5/560-stills-referring-to-ads-when-i-just-created-an-event-fred)
 - [Fix - Server side validation for checking tickets remaining is not accurate](https://trello.com/c/ZDtBrEe4/550-server-side-validation-for-checking-tickets-remaining-is-not-accurate)
 - [Fix - The remove button for a field is now working on the new ticket form](https://trello.com/c/AgyqtpDN/549-unable-to-remove-field-in-create-ticket-from-event-dashboard)
 - [Fix - Validation of start and end date validation was broken](https://trello.com/c/05y7Q30v/551-date-selection-broken-between-years)